Slight rewording of grain boundary keys
Emphasizing that the "relaxed" energies here are *supposed to be* the
true ground state energies, as we do in all of our properties.  That is,
what a Test reports as a relaxed grain boundary energy in these
properties is intended to be the true ground state energy, i.e. the
lowest energy that could be attained.